+++ internal/service/mfa/resource_fido2_policy.go
@@ -687,22 +687,25 @@ func (r *FIDO2PolicyResource) Delete(ctx context.Context, req resource.DeleteReq
 	}
 }
 
-var mfaFido2PolicyDeleteCustomError = func(p1Error model.P1Error) diag.Diagnostics {
+var mfaFido2PolicyDeleteCustomError = func(r *http.Response, p1Error *model.P1Error) diag.Diagnostics {
 	var diags diag.Diagnostics
 
-	// Undeletable default FIDO2 policy
-	if v, ok := p1Error.GetDetailsOk(); ok && v != nil && len(v) > 0 {
-		if v[0].GetCode() == "CONSTRAINT_VIOLATION" {
-			if match, _ := regexp.MatchString("cannot delete the default policy", v[0].GetMessage()); match {
+	if p1Error != nil {
+		// Undeletable default FIDO2 policy
+		if v, ok := p1Error.GetDetailsOk(); ok && v != nil && len(v) > 0 {
+			if v[0].GetCode() == "CONSTRAINT_VIOLATION" {
+				if match, _ := regexp.MatchString("cannot delete the default policy", v[0].GetMessage()); match {
 
-				diags.AddWarning("Cannot delete the default MFA FIDO2 policy", "Due to API restrictions, the provider cannot delete the default FIDO2 policy for an environment.  The policy has been removed from Terraform state but has been left in place in the PingOne service.")
+					diags.AddWarning("Cannot delete the default MFA FIDO2 policy", "Due to API restrictions, the provider cannot delete the default FIDO2 policy for an environment.  The policy has been removed from Terraform state but has been left in place in the PingOne service.")
 
-				return diags
+					return diags
+				}
 			}
 		}
 	}
 
-	return framework.CustomErrorResourceNotFoundWarning(p1Error)
+	diags.Append(framework.CustomErrorResourceNotFoundWarning(r, p1Error)...)
+	return diags
 }
 
 func (r *FIDO2PolicyResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
@@ -774,13 +777,19 @@ func (p *FIDO2PolicyResourceModel) expand(ctx context.Context) (*mfa.FIDO2Policy
 	if !mdsAuthenticatorRequirementsPlan.AllowedAuthenticatorIDs.IsNull() && !mdsAuthenticatorRequirementsPlan.AllowedAuthenticatorIDs.IsUnknown() {
 		allowedAuthenticators := make([]mfa.FIDO2PolicyMdsAuthenticatorsRequirementsAllowedAuthenticatorsInner, 0)
 
-		var allowedAuthenticatorIDsPlan []string
+		var allowedAuthenticatorIDsPlan []types.String
 		diags.Append(mdsAuthenticatorRequirementsPlan.AllowedAuthenticatorIDs.ElementsAs(ctx, &allowedAuthenticatorIDsPlan, false)...)
 		if diags.HasError() {
 			return nil, diags
 		}
 
-		for _, allowedAuthenticatorIDPlan := range allowedAuthenticatorIDsPlan {
+		allowedAuthenticatorIDs, d := framework.TFTypeStringSliceToStringSlice(allowedAuthenticatorIDsPlan, path.Root("mds_authenticators_requirements").AtName("allowed_authenticator_ids"))
+		diags.Append(d...)
+		if diags.HasError() {
+			return nil, diags
+		}
+
+		for _, allowedAuthenticatorIDPlan := range allowedAuthenticatorIDs {
 
 			allowedAuthenticator := *mfa.NewFIDO2PolicyMdsAuthenticatorsRequirementsAllowedAuthenticatorsInner(
 				allowedAuthenticatorIDPlan,