+++ internal/service/sso/resource_sign_on_policy.go
@@ -295,22 +295,25 @@ func (r *SignOnPolicyResource) Delete(ctx context.Context, req resource.DeleteRe
 	}
 }
 
-var signOnPolicyDeleteCustomError = func(p1Error model.P1Error) diag.Diagnostics {
+var signOnPolicyDeleteCustomError = func(r *http.Response, p1Error *model.P1Error) diag.Diagnostics {
 	var diags diag.Diagnostics
 
-	// Undeletable default SOP
-	if v, ok := p1Error.GetDetailsOk(); ok && v != nil && len(v) > 0 {
-		if v[0].GetCode() == "CONSTRAINT_VIOLATION" {
-			if match, _ := regexp.MatchString("delete a default policy", v[0].GetMessage()); match {
+	if p1Error != nil {
+		// Undeletable default SOP
+		if v, ok := p1Error.GetDetailsOk(); ok && v != nil && len(v) > 0 {
+			if v[0].GetCode() == "CONSTRAINT_VIOLATION" {
+				if match, _ := regexp.MatchString("delete a default policy", v[0].GetMessage()); match {
 
-				diags.AddWarning("Cannot delete the default Sign On policy", "Due to API restrictions, the provider cannot delete the default sign on policy for an environment.  The policy has been removed from Terraform state but has been left in place in the PingOne service.")
+					diags.AddWarning("Cannot delete the default Sign On policy", "Due to API restrictions, the provider cannot delete the default sign on policy for an environment.  The policy has been removed from Terraform state but has been left in place in the PingOne service.")
 
-				return diags
+					return diags
+				}
 			}
 		}
 	}
 
-	return framework.CustomErrorResourceNotFoundWarning(p1Error)
+	diags.Append(framework.CustomErrorResourceNotFoundWarning(r, p1Error)...)
+	return diags
 }
 
 func (r *SignOnPolicyResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {